health stuff

+++ Assets/Scripts/Status/HealthStatus.cs
@@ -9,7 +9,7 @@ public class HealthStatus : StatusEffect {
     public int op;
     public List<StatusEffect> stati;
 
-    private Func<float, HealthManager, float> InvokeFunc;
+    private Func<int, HealthManager, int> InvokeFunc;
 
     public enum Type {
         Hurt,
@@ -18,7 +18,7 @@ public class HealthStatus : StatusEffect {
     }
     public Type type;
 
-    public HealthStatus(HealthManager on, Type triggers, Func<float, HealthManager, float> Invoke): base(on) {
+    public HealthStatus(HealthManager on, Type triggers, Func<int, HealthManager, int> Invoke): base(on) {
         type = triggers;
         InvokeFunc = Invoke;
         if (triggers == Type.Hurt) {
@@ -37,32 +37,32 @@ public class HealthStatus : StatusEffect {
         return new HealthStatus(healthManager, type, InvokeFunc);
     }
 
-    public void Invoke(float damage) {
+    public void Invoke(int damage) {
 
     }
 }
 
 public static class HealthStatusBuilder {
     public static HealthStatus HealMultiplier(HealthManager on, float mult) {
-        return new HealthStatus(on, HealthStatus.Type.Heal, (float d, HealthManager on) => mult * d);
+        return new HealthStatus(on, HealthStatus.Type.Heal, (int d, HealthManager on) => (int)(mult * d));
     }
     public static HealthStatus HurtMultiplier(HealthManager on, float mult) {
-        return new HealthStatus(on, HealthStatus.Type.Hurt, (float d, HealthManager on) => mult * d);
+        return new HealthStatus(on, HealthStatus.Type.Hurt, (int d, HealthManager on) => (int)(mult * d));
     }
     public static HealthStatus Multiplier(HealthManager on, float mult) {
-        return new HealthStatus(on, HealthStatus.Type.Both, (float d, HealthManager on) => mult * d);
+        return new HealthStatus(on, HealthStatus.Type.Both, (int d, HealthManager on) => (int)(mult * d));
     }
-    public static HealthStatus HealModifier(HealthManager on, float mod) {
-        return new HealthStatus(on, HealthStatus.Type.Heal, (float d, HealthManager on) => d + mod);
+    public static HealthStatus HealModifier(HealthManager on, int mod) {
+        return new HealthStatus(on, HealthStatus.Type.Heal, (int d, HealthManager on) => d + mod);
     }
-    public static HealthStatus HurtModifier(HealthManager on, float mod) {
-        return new HealthStatus(on, HealthStatus.Type.Hurt, (float d, HealthManager on) => d + mod);
+    public static HealthStatus HurtModifier(HealthManager on, int mod) {
+        return new HealthStatus(on, HealthStatus.Type.Hurt, (int d, HealthManager on) => d + mod);
     }
-    public static HealthStatus Modifier(HealthManager on, float mod) {
-        return new HealthStatus(on, HealthStatus.Type.Both, (float d, HealthManager on) => d + mod);
+    public static HealthStatus Modifier(HealthManager on, int mod) {
+        return new HealthStatus(on, HealthStatus.Type.Both, (int d, HealthManager on) => d + mod);
     }
     public static HealthStatus Statifier(HealthManager on, StatusEffect status) {
         Debug.Log("aaa");
-        return new HealthStatus(on, HealthStatus.Type.Both, (float d, HealthManager on) => {status.Copy().SetManager(on); on.didItWork.Add("Yes!"); return d;});
+        return new HealthStatus(on, HealthStatus.Type.Both, (int d, HealthManager on) => {status.Copy().SetManager(on); on.didItWork.Add("Yes!"); return d;});
     }  
 }
